feat(background): set background image in its own component
